api/src/requirements.txt
PyJWT
fastapi
requests
fastapi
uvicorn
psycopg2-binary